logfiles: ownership
By running `grommunio-sync-top` the ownership of LOGFILE and LOGERRORFILE get changed to user `root`.
Because of those changes the logrotate.d-service is likely to fail.

So i would propose something like this to get a handle on this problem. It definitely should be improved with some sane Tests on all included files. Maybe a routine which first makes sure all corresponding Access-Rights are correct, change those if necessary and finally change the User of the Process?