Fixes #9
This fix adds the users primary email address to the sender address property in the MAPI message if the client is "WindowsMail" and the property is empty. Whilst I can't imagine a use case for sending as the system user over ActiveSync, I treat this as a quirk of WindowsMail and thus only modify the property if the User-Agent matches.